yelp_data_pipeline/requirements.txt
jupyter
pandas
numpy
requests
treelib
geopandas
yelpapi
google-cloud-firestore
google-cloud-secret-manager
autopep8